SNAP hardening: the reds stay together, a cancelled hue lands on the detent, and a big photo cannot take the process down
An adversarial pass over the merged engine and screen, ten findings, all
fixed here. The two that changed what a user would hear:

- Two reds, two octaves apart. TUNE was hue/360, which put the hue
  circle's seam at 0° — a 355° rose and a 5° scarlet, visually the same
  red, landed at the top and the bottom of the knob. The circle is now
  cut at rose (330°), where photos rarely sit, so every red a camera is
  likely to see (a sunset at 10°, a skin tone at 20°, a red door at
  355°) is on the same low end, and the spectrum's order is kept with
  violet high. A test holds the two reds within 0.08 of each other.

- Colours that cancel. A photo half pure red and half pure cyan is fully
  saturated, so the grey guard did not fire, but its hue vectors sum to
  nothing and atan2 of nothing is float noise: TUNE anywhere. The
  reading now carries hueStrength, the length of the mean hue vector,
  and below 0.2 the photo has no dominant hue and lands on the centre
  detent like a grey one.

Then the ones that changed what a user would measure or risk:

- Detail was measured across only and per pixel, so a photo of
  horizontal blinds read as smooth and the same scene at 160 px and
  1024 px differed six-fold in GRIT, by camera app. It is measured on a
  128-cell grid now, across and down, so the number is about the
  picture; the test holds a 128 px and a 512 px thumbnail of one scene
  within 25 % and the two stripe directions equal.

- TakePicturePreview is documented as a small bitmap, but it crosses
  from the camera app through shared memory and some camera apps fill
  it with the whole frame — a 12 MP one wants a 48 MB pixel array, and
  the OutOfMemoryError that raises is an Error, not an Exception, so the
  screen's catch let it through and the process died. The bitmap is
  shrunk to 512 px on the long side before reading (nothing Snap
  measures needs more), and the Error is caught and toasted as
  Copy.SNAP_TOO_BIG, the way TapeScreen and ChopScreen already catch it
  on their own decodes.

- SEND TO PAD read the table state and the voice state separately, so a
  tap on PLUMB followed quickly by SEND could land HORIZON's table under
  PLUMB's name; and the render effect keyed on table and macros could
  play the old photo's line under a new photo's knobs for a beat. The
  screen now keeps the line it read together with the photo and chip
  that read it (a private Line class), SEND is off while that is stale,
  and what lands is what was heard.

- A flat table typed into kit.json rendered a silent pad that nothing
  refused: SnapPatch now holds Snap.read's own refusal, and a
  wrong-length, out-of-range, flat or non-integer table from a sidecar
  is a JsonException like every other malformed recipe rather than a
  bare IllegalArgumentException — BREED and the recipe replay catch both
  already, so this is consistency, not a rescue. Engine and version are
  checked before the table is read, so a later version's file says
  "unsupported version" instead of something about its table.

- The patch keeps its own copy of the table, so a caller's later write
  to its array cannot change the sound behind the 0..255 check.

- Photo's size check multiplied in Int, so 65536 x 65536 with an empty
  array passed and indexed out of bounds later. Long arithmetic now.

Checked and found sound, by the review and by new tests: 1x1, 1xN, Nx1
and 3x700 photos read without crashing and refuse in words where their
line is flat; cycle and synthesize have no division by zero, no NaN, and
a phase step well under the table length; the HARDWARE bitmap path; the
lifecycle release of the audition voice.

+++ synth/src/main/kotlin/com/snipsnap/synth/Photo.kt
@@ -14,7 +14,9 @@ class Photo(val width: Int, val height: Int, val argb: IntArray) {
 
     init {
         require(width > 0 && height > 0) { "a photo needs a width and a height: ${width}x$height" }
-        require(argb.size == width * height) {
+        // Long arithmetic: 65536 x 65536 wraps to zero in Int and would
+        // let an empty array through to an index error later.
+        require(argb.size.toLong() == width.toLong() * height.toLong()) {
             "pixel count ${argb.size} is not ${width}x$height"
         }
     }
